Update using-a-package.json.md URL and Text
The "Learn More" text refers to "Installing npm packages locally," and refers to  a page with a video about this. It says this information can be found in "Chapter 8," linked as https://docs.npmjs.com/getting-started/installing-npm-packages-globally, however this chapter is titled "Installing NPM Packages Globally," so I believe this is the wrong chapter and link. Also, there's no video.

In this PR I've updated the text to refer to Chapter 6, "How to Update Local Packages," which seems to be the intended target due to title as well as the fact that it has a video. The link has been updated to https://docs.npmjs.com/getting-started/updating-local-packages

use chapter 4, not chapter 6